Disable testing callgraph and branch stack [2/3]
This series of patches is to make ProcessEvents return false when a
sample cannot be mapped.

In this CL, we temporarily disable testing samples with callchain and
branch stack data, as these contains various addresses that cannot be
mapped. There's a lot of data to investigate and we don't want it to
block this change, so we'll remove it from testing for now.

BUG=none
TEST=unit tests pass

Change-Id: Ic92e547a6de4b56904c21acfe4b186508880bb07
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182359