quipper: handle invalid callchain data gracefully.
Prior to this change, quipper would fail with a failed CHECK if the
callchain was invalid (if it had fewer than 2 entries, etc.). This
change makes it return false to the caller instead of failing a CHECK.

It also adds a test case with a callchain of size 1.

TEST=unit tests pass.
BUG=none

Change-Id: I0f5065149b8147e88678ec726f83a92f09fc4a4f
Reviewed-on: https://chromium-review.googlesource.com/182801
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Queue: Ahmad Sharif <asharif@chromium.org>
Tested-by: Ahmad Sharif <asharif@chromium.org>