Clean up object ownership
The session_manager has, over time, grown a funky dependency graph.
In particular, SessionManagerService owns a lot of things that
then point back to it, sometimes in an opaque way. Cleaning up this
ownership graph makes it easier to extract the code that listens
for incoming signals, which I need to do to move to MessageLoopForIO.

BUG=chromium:330882
TEST=unit
TEST=suite:smoke
TEST=all the Ownership autotests

Change-Id: I5fc3fdb946f6c709e75d223e00f284d4e04b8492
Reviewed-on: https://chromium-review.googlesource.com/183721
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>