Remove LOG(FATAL) in AllowGracefulExit()
I forgot that there are tests and developers that do, indeed, need the
code pathway in the session_manager that causes the message loop to
run until it's explicitly killed. These are tests that want to tear
down Chrome and manage the browser or some other process that needs
the screen manually.

This change removes a LOG(FATAL) I'd added to that code pathway because I
thought it was in fact unneeded.

BUG=chromium:335302
TEST=unit
TEST=video_VideoDecodeAccelerator

Change-Id: I47a11080278d7cc51e9e83601117752cea754ff1
Reviewed-on: https://chromium-review.googlesource.com/182929
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>