Change factory-splash job to start on starting boot-services.
When performing a powerwash on a device, the startup job launches the
chromeos-boot-alert script which emits the boot-alert-request event.
This event is emitted in a blocking way, which will make the caller
(startup) to wait until the boot-splash job is stopped.

On the other hand, CL:181323 included a new factory-splash job on
test images only, that depends on both boot-splash and startup
"stopping" event. This means that this new job will block the
emitted event until startup is stopped, creating a deadlock.

This patch makes the factory-splash.conf job to start on
boot-services starting instead, which breaks the deadlock.

BUG=chromium:336500
TEST=platform_Powerwash works on a test image and the factory splash
screen is shown when /usr/local/factory/enabled is present.

Change-Id: Iafd1848bf78acb5fc770a3f1aca3cf7ce1c48154
Reviewed-on: https://chromium-review.googlesource.com/183363
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>