Move signal watching into objects owned by SessionManagerService
The session_manager registers handlers for a number of signals.
Some (termination signals and SIGCHLD) have handlers that write to a pipe
so that the arrival of the signal can be handled on the main event loop in
a normal execution environment.

Until now, this code has all lived inside SessionManagerService. Breaking it
out reduces the complexity of this class, makes testing easier and, most
importantly, moves to a model that is more easily adapted to MessageLoopForIO.

BUG=chromium:330882
TEST=unit
TEST=suite:smoke

Change-Id: Ic51dd448929463ae4f05a4a2efc488c2915b9150
Reviewed-on: https://chromium-review.googlesource.com/184501
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>