*.proto: define key parameters and extend vault_keyset
This change adds the requisite key properties and data
required for key add/auth/update to the proto bufs. The
changes to SerializeVaultKeyset are the first steps towards
better Web-integration of the user-agent principals.  The
KeyParameters are a move to a easier structure for passing
key objects through DBus.  (The alternative is to define
a structure in the dbus interface which then requires custom
wrapping on the Chrome-side (see cros-disks).)

BUG=chromium:316189
TEST=compile test. Protos are optional and unreferenced.

Change-Id: I8c4de48d393556721d6943f0e1b0b0faf92400ba
Reviewed-on: https://chromium-review.googlesource.com/185597
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>