Call 'LockScreen' method instead of sending signal
When the session manager wants to tell the browser to show the
lock screen, we broadcast a signal over DBus due to ancient
stupidity. The right way to do this is to have a method on
Chrome called 'LockScreen', and then call it from the session
manager. So that's what this code does, now.

BUG=chromium:341925
TEST=desktopui_ScreenLocker, also manual screen locking.

Change-Id: I722e336be304a894d79aff59b7a376bdef969ef5
Reviewed-on: https://chromium-review.googlesource.com/185412
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>