Stop sharing chaps::TokenManagerClient.
Instances of Mount can be used concurrently by multiple threads.
chaps::TokenManagerClient is not thread-safe so it should not be a
member of Mount.

This fixes a crash when InsertPkcs11Token and RemovePkcs11Token race to
initialize the dbus connection.

BUG=chromium:262654
TEST=unit, manual
I cannot repro the crash in this bug so we'll need to watch for it to
make sure this is fixed.

Change-Id: I01ce73e66608d8c1b60211c7271fbd006674a226
Reviewed-on: https://chromium-review.googlesource.com/186293
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>