MountEx, key privileges, KeyData implementation
This change implements MountEx which uses the new interface
for talking with Cryptohomed.  Additionally, the privilege
checks are added to AddKeyset, Migrate, and MountEx (but not
yet to (Async)Mount).  Automated testing is still outstanding
 as well as manual testing of key privileges, but the primary
pieces are in place and are behaving as expected.

BUG=chromium:316189
TEST=manual testing with mount_ex and return the sanitized username
     all _existing_ unittests still pass
     new unittests pass; need migrate permission testing, but lack Migrate unittests.
Change-Id: I4552c28caf2bfb833a15492ec2b61ff49cc2760e
Reviewed-on: https://chromium-review.googlesource.com/187200
Reviewed-by: Will Drewry <wad@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>