shill: reload_wifi: New PCI device helper script
This script is UDEV helper function that can help restart a PCI
based wireless LAN device.  It does so by removing the PCI device
from the bus, using the EC to cycle power on the device, then
performing a bus re-enumeration of the parent bus of the device.

BUG=chrome-os-partner:25275
TEST=Manual: Force restart by modifying the driver to output an event,
and ensure the device disappears and reappears as a new phy.

Change-Id: Ife0b05f8f545623d15d828610f180bb79669f926
Reviewed-on: https://chromium-review.googlesource.com/187474
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>