Stop watching term signal FD once it's handled.
To handle shut down, the session manager registers a signal handler
that writes to a pipe when a termination signal comes in. The old
glib-based code that watched the pipe used to de-register itself once
it ran. We've had at least one bug crop up where the session_manager
just spins sending SIGTERM to the browser for a long time, and it's
probably because the new code doesn't do the same -- and also doesn't
read the data from the pipe.

BUG=chromium:346511
TEST=login_LoginSucces, and check /var/log/messages to be sure the session_manager isn't SIGTERMing the browser hundreds of times.

Change-Id: If13061767c5913481f28ef4d9b0fe2569b2b0bbc
Reviewed-on: https://chromium-review.googlesource.com/187728
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>