Fix crash in some calls to RestartJob
Not setting a GError* to NULL initially can cause a malformed
error message to be sent back from g_shell_parse_argv.
This can then propagate out to the chrome dbus bindings and cause
a crash when it tries to send an error with this wonky payload.

BUG=None
TEST=login_LogoutProcessCleanup, unit tests

Change-Id: Ic262cbefdbb18ffc63b5002313b17333745483aa
Reviewed-on: https://chromium-review.googlesource.com/187942
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>