PolicyManager: Add a new State class grouping the state providers.
The policy implementations need access to the providers to discover
the variables exposed by them. Instead of having global variables to
access each providers, we add a new State interface that groups all
the providers in the same way a provider groups all the variables.

This interface class allows to create a mock/fake class for testing
the policies.

Other minor fixes to the fakes are included in this patch.

BUG=chromium:338591
TEST=Simple unit test passes.

Change-Id: I7fe46dfc8416ee39ace3290628b7bae440213b29
Reviewed-on: https://chromium-review.googlesource.com/187705
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>