shill: OpenVPNDriver: Accept and use Token parameter
The OpenVPN.Token property is similar to the OpenVPN.OTP property
in that it is a one-time-use credential that can be used during
Static Challenge authentication.  The semantics of how this
credential is presented is slightly different, and it does not
require a password parameter.

BUG=chromium:347298
TEST=Unit tests

Change-Id: I9ee8160b40da8b71c1facc2bdfe9453b2a4f3c6e
Reviewed-on: https://chromium-review.googlesource.com/188165
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>