shill test scripts: make scripts more consistent
For some reason, the invocation of connect-service,
disconnect-service, and set-service-property was different.
connect-serivce wanted the Name property of the Service,
while disconnect-service and set-service-propety wanted
something else.

Change this, so that all three scripts work with a Service
Name. Also, set the executable bit on set-service-property.

BUG=None
TEST=manual (see below)

1. set up ap
2. let shill find it
3. connect-service <wifi-ssid>
4. set-service-property <wifi-ssid> AutoConnect false
5. disconnect-service <wifi-ssid>
6. wait 30 seconds, check that the wifi service is still
   disconnected

Change-Id: I00cae9e2b89cdbc6746500c4d0b5c769dc90970e
Reviewed-on: https://chromium-review.googlesource.com/188388
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>