Support devices that expose a MBIM interface.
This CL modifies UsbModemSwitchOperation to first check if the device
has a USB configuration that exposes a MBIM interface. If a MBIM
configuration exists, the switch operation completes by selecting that
configuration. Otherwise, the switch operation proceeds the same way as
it used to be by sending special USB messages.

BUG=chromium:349256
TEST=Tested the following:
1. Build and run unit tests.
2. Verify that mist successfully switches supported non-MBIM dongles
   from the mass storage mode to the modem mode.
3. Verify that mist successfully selects the MBIM configuration for
   supported MBIM dongles.

Change-Id: I288517f9f153ec2dc9a8ece22dcffacfc559baae
Reviewed-on: https://chromium-review.googlesource.com/188988
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>