Pass '--gpu-sandbox-failures-nonfatal' flag.
The current code in Chrome already considers GPU sandbox failures
nonfatal, so this is not degrading security. By passing this flag
we'll be able to tighten the code in Chrome, without having to worry
about breaking Chrome OS. Once the Chrome code rolls into Chrome OS,
we can test the *removal* of this flag with trybots on all boards.

The other way around is not possible since it's not possible to have
Chrome OS tryjobs of Chrome CLs.

BUG=chromium:327824
TEST=Trybots.

Change-Id: Id1394ee04ec45cf55bb65ae73544ef6753ec4e3d
Reviewed-on: https://chromium-review.googlesource.com/188814
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Julien Tinnes <jln@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>