session_manager_setup.sh: increase fd ulimit at startup
The normal place to configure ulimits is in /etc/security/limits.conf and
/etc/security/limits.d. Unfortunately this settings are only honored when
authenticated through PAM. In other words, for system startup scripts
those files are not consulted.

Because of this, the way to increase these limits for chrome (and all
subprocesses) will have to be done either in the upstart script or in
the actual launch script. I went with the latter since we're alreayd
modifying limits in this script as it is.

BUG=chromium:251385
TEST=boot, start a session, launch a shell from under chrome, check that
"ulimit -n" output as cronos user is 2048. I will also open a bug to
track an autotest for this.

Change-Id: I5cf1e5707a51cff2aec379db090fa70e2e0a7ab9
Reviewed-on: https://chromium-review.googlesource.com/188893
Tested-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>