power: Delete old keep_backlight_on_for_audio pref.
Delete an old keep_backlight_on_for_audio powerd pref. This
was added before powerd was updated to check for HDMI audio
and it doesn't appear to be set anywhere.

Also log a message when audio is being sent to the display
at the point where the screen is dimmed to make it more
obvious why the screen won't be turned off.

BUG=chrome-os-partner:13240
TEST=deleted old tests; also manually tested that screen is
     still kept on for HDMI audio and that new message is
     logged

Change-Id: Ie02d1e883ad38e9fa3a16c161ece85b97af6a932
Reviewed-on: https://chromium-review.googlesource.com/190310
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>