Set US as a strict regdomain
Going forward, it is impractical for many radio vendors to support
channel 12 and 13 to FCC regulations.  As such, we should treat
the "US" regdomain as a strict regdomain so as to prohibit passive
scanning on these channels.

References:
http://transition.fcc.gov/bureaus/oet/ea/presentations/files/oct13/70-Software-Configurations-and-DFS-issues-.pdf
http://www.hctamerica.com/blog-radio_9.html

BUG=chromium:353773
TEST=Boot system with no default networks (so it doesn't associate
at boot) with "region=us" in the VPD and do "iw reg get".  It should
report "US" as the regdomain.

Change-Id: I5e82df6adad1f8334d649afb8b4da9298449e949
Reviewed-on: https://chromium-review.googlesource.com/190760
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>