Make setup of regulatory domain more reliable
Initially when regulatory_domain.conf was created, we worked
around a startup issue by calling "iw reg set" twice.  It
turns out that this was a race condition between wireless
driver startup and setting of the regulatory domain.  Add a
retry loop to the regulatory setup script so that we wait
for a reasonable amount of time for the wireless LAN driver
to initialize before setting the regulatory domain.

BUG=chromium:353773
TEST=On a system where the wireless driver takes a while to load,
but for which a strict regdomain is configured, test the output of
both "iw reg get", and "iw list" to make sure that the appropriate
channels are disabled.

Change-Id: I987d441c929a0f47366c410a9c2a449509f0dc59
Reviewed-on: https://chromium-review.googlesource.com/190761
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>