AddKeyEx/AddKeyset: add label clobber support
Allow a key to overwrite another key if they share
a label.  This has a direct impact on key migration
with labeled keys as well as on the AddKeyEx call
which supports the |clobber_if_needed| argument.

As part of this, VaultKeyset was annotated with its
loaded index, optionally, and all AddKeyset callsites
were updated.

TEST=build, unit, autotests (*Cryptohome*)
BUG=chromium:316189

Change-Id: Ide62b3b4243d7e86574c4bedf8b1222c3cc381a7
Reviewed-on: https://chromium-review.googlesource.com/191023
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>