Remove sync() from unit tests and clean up Mount tests.
The initial intention of this CL was to mock out calls to sync() in unit
tests.  In the process of fixing this I found that a lot of tests using
the MountTest fixture had a use-after-free on their local mocks.  In
fixing this I found a few other spots in the Mount tests that needed
cleaning up including real dbus calls to chaps instead of using a mock
chaps client.  So this CL includes the following cleanups:
- All calls to sync() are mocked
- All calls to a chaps client are mocked
- The MountTest fixture automatically configures a Mount class with the
  right (nice) mocks in a default state and tears them down correctly.

BUG=chromium:350959
TEST=Ran the unit tests with an CHECK in Platform::Sync to ensure it is
not called.

Change-Id: Icb5e61489d4aea75111bd18cff7def5b85089551
Reviewed-on: https://chromium-review.googlesource.com/191546
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>