service.cc: Move AsyncMount, AsyncMountPublic to the mount_thread
This change moves all AsyncMount and AsyncMountPublic logic to
the mount_thread_ allowing a DBus reply to be processed as quickly
as a new sequence number can be generated by a new MountTaskMount.

This also fixes an oversight in ::Mount where failure left a Mount
object in the mounts list.  This can be broken out if needed but I
don't believe we have many (any?) consumers of Mount or MountPublic,
just the async version.   This also pointed out a minor issue in a
test which was fixed (ignoring a failed Mount call).

TEST=manual testing works. unittests still normal (poor coverage in this area).
     autotest all cryptohome pass -- suite:bvt runnign now.
BUG=chromium:358781

Change-Id: Iadd3975ec283bf1944255bf6382172eb44947fa0
Reviewed-on: https://chromium-review.googlesource.com/192621
Reviewed-by: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>